requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.4
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.0
frozenlist==1.4.0
httplib2==0.20.2
importlib-metadata==4.6.4
iniconfig==2.0.0
isort==5.13.2
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
logilab-common==1.8.2
mccabe==0.7.0
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==0.4.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.0.3
pyparsing==2.4.7
pytest==7.4.3
pytest-dotenv==0.5.2
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.0
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
typing-extensions==3.10.0.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
wrapt==1.13.3
yarl==1.9.4
zipp==1.0.0
